+++ x/evm/types/genesis_test.go
@@ -56,23 +56,6 @@ func (suite *GenesisTestSuite) TestValidateGenesisAccount() {
 			"empty account address bytes",
 			GenesisAccount{
 				Address: ethcmn.Address{}.String(),
-				Balance: sdk.OneInt(),
-			},
-			false,
-		},
-		{
-			"empty account balance",
-			GenesisAccount{
-				Address: suite.address,
-				Balance: sdk.Int{},
-			},
-			false,
-		},
-		{
-			"negative account balance",
-			GenesisAccount{
-				Address: suite.address,
-				Balance: sdk.NewInt(-1),
 			},
 			false,
 		},
@@ -80,7 +63,6 @@ func (suite *GenesisTestSuite) TestValidateGenesisAccount() {
 			"empty code bytes",
 			GenesisAccount{
 				Address: suite.address,
-				Balance: sdk.OneInt(),
 				Code:    "",
 			},
 			false,
@@ -99,7 +81,6 @@ func (suite *GenesisTestSuite) TestValidateGenesisAccount() {
 }
 
 func (suite *GenesisTestSuite) TestValidateGenesis() {
-
 	testCases := []struct {
 		name     string
 		genState *GenesisState
@@ -116,7 +97,6 @@ func (suite *GenesisTestSuite) TestValidateGenesis() {
 				Accounts: []GenesisAccount{
 					{
 						Address: suite.address,
-						Balance: sdk.OneInt(),
 						Code:    suite.code,
 						Storage: Storage{
 							{Key: suite.hash.String()},
@@ -168,7 +148,6 @@ func (suite *GenesisTestSuite) TestValidateGenesis() {
 				Accounts: []GenesisAccount{
 					{
 						Address: suite.address,
-						Balance: sdk.OneInt(),
 						Code:    suite.code,
 						Storage: Storage{
 							NewState(suite.hash, suite.hash),
@@ -176,7 +155,6 @@ func (suite *GenesisTestSuite) TestValidateGenesis() {
 					},
 					{
 						Address: suite.address,
-						Balance: sdk.OneInt(),
 						Code:    suite.code,
 						Storage: Storage{
 							NewState(suite.hash, suite.hash),
@@ -192,7 +170,6 @@ func (suite *GenesisTestSuite) TestValidateGenesis() {
 				Accounts: []GenesisAccount{
 					{
 						Address: suite.address,
-						Balance: sdk.OneInt(),
 						Code:    suite.code,
 						Storage: Storage{
 							{Key: suite.hash.String()},
@@ -242,7 +219,6 @@ func (suite *GenesisTestSuite) TestValidateGenesis() {
 				Accounts: []GenesisAccount{
 					{
 						Address: suite.address,
-						Balance: sdk.OneInt(),
 						Code:    suite.code,
 						Storage: Storage{
 							{Key: suite.hash.String()},