+++ x/evm/types/storage.go
@@ -5,6 +5,8 @@ import (
 
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 
+	ethermint "github.com/cosmos/ethermint/types"
+
 	ethcmn "github.com/ethereum/go-ethereum/common"
 )
 
@@ -49,7 +51,7 @@ func (s Storage) Copy() Storage {
 
 // Validate performs a basic validation of the State fields.
 func (s State) Validate() error {
-	if IsEmptyHash(s.Key) {
+	if ethermint.IsEmptyHash(s.Key) {
 		return sdkerrors.Wrap(ErrInvalidState, "state key hash cannot be empty")
 	}
 	// NOTE: state value can be empty